requirements.txt
boto3
botocore
PyJWT
chalice
